requirements.txt
tabulate
discord.py
SQLAlchemy